requirements.txt
flask
torch
torchvision
efficientnet_pytorch
Pillow
reportlab
gunicorn
